requirements.txt
fastapi
pydantic
qdrant-client
google-generativeai
pypdf
